Insert screenshot
This screenshot shows all of the functions of the app. It updates the stock after the user "purchases" items, it alerts the user with the price of the items they are buying, it alerts the user that they may not make a purchase that surpasses the stock quantity, and it ends the query connection when the user answers "no" to "Would you like to purchase an item?".